Add config header
Use a config header bed osan zconf.h.cmakein.

There is already a zconf.h in the repo, that is removed by this commit.
It conflicts and takes precedence over the the generated zconf.h, so it
has to go. The CMake file also has logic to move away the shipped
zconf.h file. Seems quite strange TBH. Why is it there?